03. Next Components Server and Client | 18. Exploring Which Components Should Be Client and Which Should Be Server
