+++ app/src/app/restaurant/[slug]/components/RestaurantReservationCard.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 export default function RestaurantReservationCard() {
   return (
     <div className="fixed w-[15%] bg-white rounded p-3 shadow">
